+++ Gem/Code/Source/RHI/AlphaToCoverageExampleComponent.cpp
@@ -393,7 +393,7 @@ namespace AtomSampleViewer
                 RHI::DrawItem drawItem;
                 drawItem.m_arguments = drawIndexed;
                 drawItem.m_pipelineState = m_pipelineStates[typeIndex].get();
-                drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                 drawItem.m_shaderResourceGroups = shaderResourceGroups;
 
                 const RHI::IndexBufferView indexBufferView =
@@ -406,7 +406,7 @@ namespace AtomSampleViewer
                 drawItem.m_indexBufferView = &indexBufferView;
 
                 AZStd::array<AZ::RHI::StreamBufferView, 2>& streamBufferViews = m_rectangleStreamBufferViews;
-                drawItem.m_streamBufferViewCount = static_cast<uint32_t>(streamBufferViews.size());
+                drawItem.m_streamBufferViewCount = static_cast<uint8_t>(streamBufferViews.size());
                 drawItem.m_streamBufferViews = streamBufferViews.data();
 
                 // Submit the rectangle draw item.
